requirements.txt
click==8.0.1
colorama==0.4.4
install==1.3.4
Pillow==8.3.1
